Milestone 1 - Improved Order Processing
Yes, Milestone 1 was completed before the MVP. Blame Xander not getting me the information I need to verify for Stripe :p

# Employee Access Page
Access page is at /staff-portal currently only contains the orders page, but has the framework for further accessible pages. The sidebar to access these is hidden at this time because of that.
# Orders Page
Includes basic login system and authorisation/authentication. Currently users can only be given permission to access the orders page by directly editing the JSON properties attached to their account in Supabase.
# 404 Page
Basic catch-all in the BrowserRouter for a 404 page, self-explanatory.