Fixes #19 RM Dispatch Warnings on Fulfilled Orders
Warnings have been reworked to only show on unfulfilled orders, and a few bits of code have been reorganised to be more legible with the multiple options of display modes they have to allow for all the dynamic display of colours and warnings etc.

Also made aware of a new issue, #28, while working on this fix.