Product Image Optimisation Overhaul
Images are automatically pre-transformed and compressed to give faster page loading speeds and lower Storage Egress costs, allowing us to stay under the new quota with the paid plan as the website grows.

Also fixes various bugs with images not appearing square in mobile view etc. with a new SquareImageBox component which can be reused site-wide to save time.