Session Notifications
On session start a notification can be displayed based on the `session_notif` site setting. A start time and end time can be set to only display this notification within a set time frame, allowing for "holiday periods" where customers are warned when orders may not ship until after the period.