Disable products when out of stock #33
Products display a message when they are out of stock. The button is now properly disabled and greyed out, rather than just appearing to not be working properly. This is purely UX and doesn't serve a technical purpose, but makes it much clearer to customers why they may not be able to buy something.

These messages are editable from the site settings on Supabase, so the code doesn't need to be recompiled if the message is changed.

Also swapped all instances of `var` to `let`