PARTIALLY COMPLETE Milestone 2.5 - Report View
* Created Page

* Reorganised src/staff folder

* Created generic <objectListItem/>

Used to display generic items in a list, this component is based off of the staff order view. It'll also be used for reports in order to keep styling consistent between the two views.

* Permission checking

* <Page/> Component

Instead of repeating code for every page, each page now uses a <Page/> component which sets up a basic framework used by every page, such as header, footer, etc.

* <AuthenticatedPage/> Component

A variation of the page component which requires that the user attempting to access the page has permission to access it.

* Fetch Report Data

Fetch the report data and display it. Also includes a button which will allow those with permission to create new reports, functionality not added.

* Additional <objectListItem/> dropdown control

Dropdown toggle can now be controlled externally. The internal toggles can also be hidden independently of this, allowing for more granular control of the component in certain situations.

This will be useful for the "Create Report" button on the reports page.

* Create new report functionality

* Outlined Report Viewer

* Created report summary `objectListItem`

Resolves #129

* Delete Reports

* Fetch Report Data

* Editable Text Box

Not exactly as described in #128, but performs the same job.

Resolves #128

* Editable Text Box

Not exactly as described in #128, but performs the same job.

Resolves #128

* MDXEditor

Added MDXEditor Package, a powerful markdown editor package which rules out the need for the editableText component from previous commit.

Resolves #111

* Title Page

Resolves #119
Resolves #130

* Page Separator

Just using a styled <hr>

Resolves #110

* Fixed Editable Title Bug

Title was editable without permission, this has been amended.

* Added Table Support to Markdown Editor

Removed use of <Markdown> component in favour of simply using the readOnly mode of <MDXEditor> since things seem to just work out of the box with that and I can see no reason not to.

Tables are now supported in uses of <MDXEditorAuth>, which are used in reports.

Resolves #113

* Changed colour of title

Title now uses a more muted bone-white colour instead of the stark white it had before.

* Budget Usage Report Page

Added budget usage report page to report generator. Also tidies up a lot of the logic around saving data to reports, and various styling changes.

* Chart & Graph Components

Utilising Recharts to render charts and graphs.

Resolves #112

* Disabled Table Edit On Mobile

Disabled table edit mode on mobile to allow the budget tables to fit properly on the screen for users with edit permissions.

* Employee Hours Report Page

Resolves #121

* View Mode Toggle

Resolves #131

* Activity Report Page

Resolves #116, #122, #117

* Pre-created report component templates

Created template files for the remaining pages needed to complete the report. Hoping to get this done by the 25th so I don't have to make a report the old way again :)

* Plan Moving Forward Report Page

Resolves #127

* Attainment of Month's Plan Report Page

Resolves #126

* Issue Tracking Report Page

Resolves #123