Report View MVP (#134)
* Removed Vulnerabilities Report Page

Vulnerabilities will be handled through GitHub, either as issues if they're non-sensitive, or by the GitHub security tab, which won't be accessible by those viewing the reports unless they're an admin on the repo (I am the only admin currently)

* Site Analytics Text Box

Added a text box to the site analytics section as a placeholder for the real analytics.

* Fixed Clockify Data

Clockify data didn't include the last day of the period. It does now.