Report View (#139)
Completed report view from #134.

Adds an interactable report page which displays data and analytics about a given period. Most useful for monthly reports, although due to its flexibility it can also be used for reports of other lengths too (weekly, for example).

Reports are editable using a simple UI on the website, which is only displayed to those with permission to view it.

https://thisshopissogay.com/staff/reports

---

Also tidies up many of the types and duplicated code between Netlify functions and the client side React code using a `shared` directory at the root of the project.