Stripe Local Currency
Customers should now be charged in their local currency through Stripe.

Currently having an issue where the price calculated when generating the checkout session is different to that calculated client-side, so this will be the next thing to fix before ensuring that `purchase` and `refund` GA4 event triggers are functional in foreign currencies.