Generic Validation Methods
Schemas are generated dynamically by a script, matching all of the defined types in `@shared/types`. These schemas can be used in a generic way to validate that data is in the form expected before using it in. This is a huge deal because it means I can quickly and easily check the shape of data on the fly.