Update All Stripe Products Script
Created a script that can be used to update the products on Stripe to match Supabase. This updates exchange rates, creates new products where they don't yet exist on Stripe, and deactivates old products.

This means the checkout will be able to load faster because much of the work is done in advance and cached for later. So when users go to the checkout page, they won't have to wait for all of the products in their cart to update on Stripe (which can introduce a few seconds of latency).