Show more people's hours on reports
Added the ability to show more people's hours on reports (instead of just Finn and I). There isn't a GUI way of doing this yet, it's just now possible by manually editing the report in the backend, but that's all I need for now to add Soren to the report for this month.