Generic Ticker Component
Working on creating a ticker component to replace the very similar code that exists in multiple places to adjust the quantity of a product in the basket, as well as selecting the page number in the product list.

This is incomplete but will reduce code duplication and code quantity in general.